+++ src/datasources/results/components/variable-editors/ResultsVariableQueryEditor.tsx
@@ -19,6 +19,7 @@ import {
 import { ResultsDataSource } from 'datasources/results/ResultsDataSource';
 import { StepsQueryBuilderWrapper } from '../query-builders/steps-querybuilder-wrapper/StepsQueryBuilderWrapper';
 import { TAKE_LIMIT } from 'datasources/results/constants/QuerySteps.constants';
+import { FloatingError } from 'core/errors';
 
 type Props = QueryEditorProps<ResultsDataSource, ResultsQuery, ResultsDataSourceOptions>;
 
@@ -28,6 +29,7 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
   const [isQueryBuilderDisabled, disableStepsQueryBuilder] = useState<boolean>(true);
   const [stepsRecordCountInvalidMessage, setStepsRecordCountInvalidMessage] = useState<string>('');
   const [resultsRecordCountInvalidMessage, setResultsRecordCountInvalidMessage] = useState<string>('');
+  const [isProductSelectionInStepsValid, setIsProductSelectionInStepsValid] = useState(true);
   const queryResultsquery = query as ResultsVariableQuery;
   const stepsVariableQuery = query as StepsVariableQuery;
   const queryResultsDataSource = useRef(datasource.queryResultsDataSource);
@@ -120,6 +122,7 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
   }
 
   const onProductNameChangesinSteps = (productNames: Array<SelectableValue<string>>) => {
+    setIsProductSelectionInStepsValid(productNames.length > 0);
     onChange({ ...stepsVariableQuery, partNumberQueryInSteps: productNames.map(product => product.value as string) } as StepsVariableQuery );
   }
 
@@ -198,7 +201,12 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
       )}
       {query.queryType === QueryType.Steps && (
         <>
-          <InlineField label="Product (part number)" labelWidth={26} tooltip={tooltips.productName}>
+          <InlineField
+            label="Product (part number)"
+            labelWidth={26}
+            tooltip={tooltips.productName}
+            invalid={!isProductSelectionInStepsValid}
+            error="You must select at least one product in this field.">
             <MultiSelect
               maxVisibleValues={5}
               width={65}
@@ -240,6 +248,7 @@ export function ResultsVariableQueryEditor({ query, onChange, datasource }: Prop
           </InlineField>
         </>
       )}
+      <FloatingError message={queryResultsDataSource.current.errorTitle} innerMessage={queryResultsDataSource.current.errorDescription} severity='warning'/>
     </>
   );
 }